+++ common/lib/src/isolate/child/http_target_discovery_isolate.dart
@@ -21,7 +21,7 @@ class HttpTargetTask {
 @internal
 Future<void> setupHttpTargetDiscoveryIsolate(
   Stream<SendToIsolateData<IsolateTask<HttpTargetTask>>> receiveFromMain,
-  void Function(IsolateTaskResult<Device?>) sendToMain,
+  void Function(IsolateTaskResult<Device>) sendToMain,
   InitialData initialData,
 ) async {
   await setupChildIsolateHelper(
@@ -30,12 +30,22 @@ Future<void> setupHttpTargetDiscoveryIsolate(
     sendToMain: sendToMain,
     initialData: initialData,
     handler: (ref, task) async {
+      Object? error;
       final device = await ref.read(httpTargetDiscoveryProvider).discover(
             ip: task.data.ip,
             port: task.data.port,
             https: task.data.https,
+            onError: (url, e) => error = e,
           );
-      sendToMain(IsolateTaskResult(
+
+      if (error != null || device == null) {
+        return sendToMain(IsolateTaskErrorResult(
+          id: task.id,
+          error: error?.toString() ?? 'Unknown error',
+        ));
+      }
+
+      sendToMain(IsolateTaskSuccessResult(
         id: task.id,
         data: device,
       ));