+++ common/lib/src/isolate/parent/parent_isolate_provider.dart
@@ -24,7 +24,7 @@ const _uploadIsolateCount = 2;
 class ParentIsolateState with ParentIsolateStateMappable {
   final SyncState syncState;
   final IsolateConnector<IsolateTaskStreamResult<Device>, SendToIsolateData<IsolateTask<HttpScanTask>>>? httpScanDiscovery;
-  final IsolateConnector<IsolateTaskResult<Device?>, SendToIsolateData<IsolateTask<HttpTargetTask>>>? httpTargetDiscovery;
+  final IsolateConnector<IsolateTaskResult<Device>, SendToIsolateData<IsolateTask<HttpTargetTask>>>? httpTargetDiscovery;
   final IsolateConnector<Device, SendToIsolateData<MulticastAnnouncementTask>>? multicastDiscovery;
   final List<IsolateConnector<IsolateTaskStreamResult<double>, SendToIsolateData<IsolateTask<BaseHttpUploadTask>>>> httpUpload;
   int get uploadIsolateCount => httpUpload.length;
@@ -86,7 +86,7 @@ class IsolateSetupAction extends AsyncReduxAction<IsolateController, ParentIsola
       ),
     );
 
-    final httpTargetDiscovery = await startIsolate<IsolateTaskResult<Device?>, SendToIsolateData<IsolateTask<HttpTargetTask>>, InitialData>(
+    final httpTargetDiscovery = await startIsolate<IsolateTaskResult<Device>, SendToIsolateData<IsolateTask<HttpTargetTask>>, InitialData>(
       task: setupHttpTargetDiscoveryIsolate,
       param: InitialData(
         syncState: state.syncState,